rename file names, delete unused hooks
