+++ src/hooks/index.js
@@ -1,13 +1,4 @@
-import useArticlesInCategory from "./useArticlesInCategory";
-import useCategoryMeta from "./useCategoryMeta";
-import useAllCategoriesMeta from "./useAllCategoriesMeta";
 import useLocalStorage from "./useLocalStorage";
 import useRating from "./useRating";
 
-export {
-  useArticlesInCategory,
-  useCategoryMeta,
-  useAllCategoriesMeta,
-  useLocalStorage,
-  useRating,
-};
+export { useLocalStorage, useRating };
