--- src/hooks/useAllCategoriesMeta.js
@@ -1,7 +1,0 @@
-import categories from "../configs/categories.json";
-
-const useAllCategoriesMeta = () => {
-  return categories;
-};
-
-export default useAllCategoriesMeta;
